[tui] Optimize curses library for smaller executables
